Added a separate package BayesianNeuralPDE.jl
