+++ lib/BayesianNeuralPDE/src/BayesianNeuralPDE.jl
@@ -1,0 +1,11 @@
+module BayesianNeuralPDE
+
+include("advancedHMC_MCMC.jl")
+include("BPINN_ode.jl")
+include("PDE_BPINN.jl")
+include("pinn_types.jl")
+
+export BNNODE, ahmc_bayesian_pinn_ode, ahmc_bayesian_pinn_pde
+export BPINNsolution, BayesianPINN
+
+end
